requirements.txt
fastapi[standard]
sqlalchemy
bcrypt
psycopg[binary]
redis
python-dotenv
pyjwt
pydantic